task 12 done && code refactored
