+++ js/presentation/main.js
@@ -1,9 +1,7 @@
 import {initPictureList} from './main-screen/components/picture-list-fragment.js';
 import {initUploadingFeature} from './upload-picture/upload-picture.js';
 
-const init = () => {
-  initPictureList();
-  initUploadingFeature();
-};
 
-init();
+initPictureList();
+initUploadingFeature();
+
