+++ js/presentation/upload-picture/upload-picture.js
@@ -1,11 +1,10 @@
 import {isEscKey} from '../utils/utils.js';
-import {
-  areHashtagsCorrect,
-  hasNoHashtagDuplicates,
-  areHashtagsRightAmount, isCommentCorrect, MAX_COMMENT_SIZE, ZOOM_STEP, ZOOM_MIN, EFFECTS, ZOOM_MAX, uploadPicture
-} from './upload-picture-viewmodel.js';
+import {initScaleSettings, removeScaleListeners} from './components/scale-settings.js';
+import {initEffectsSettings, removeEffectsListeners} from './components/effects-settings.js';
+import {initPristineValidations, isPristineValid, removePristine} from './components/validation-fields.js';
+import {uploadPicture} from './upload-picture-viewmodel.js';
 
-const documentBody = document.querySelector('body');
+const documentBodyElem = document.querySelector('body');
 
 const uploadImageElem = document.querySelector('#upload-file');
 const overlayImageElem = document.querySelector('.img-upload__overlay');
@@ -16,102 +15,28 @@ const hashtagInputElem = formElem.querySelector('.text__hashtags');
 const commentInputElem = formElem.querySelector('.text__description');
 const submitButton = formElem.querySelector('.img-upload__submit');
 
-const previewImageElem = overlayImageElem.querySelector('.img-upload__preview');
+const successfulSubmissionElem = document.querySelector('#success').content.querySelector('.success');
+const errorSubmissionElem = document.querySelector('#error').content.querySelector('.error');
+const bnSuccessElem = successfulSubmissionElem.querySelector('.success__button');
+const bnErrorElem = errorSubmissionElem.querySelector('.error__button');
 
-const bnZoomOutElem = overlayImageElem.querySelector('.scale__control--smaller');
-const bnZoomInElem = overlayImageElem.querySelector('.scale__control--bigger');
-const scaleControlElem = overlayImageElem.querySelector('.scale__control--value');
+const previewElem = document.querySelector('.img-upload__preview').querySelector('img');
+const fileChooserElem = document.querySelector('.img-upload__start input[type=file]');
 
-const effectsListElem = overlayImageElem.querySelector('.effects__list');
-const sliderElem = overlayImageElem.querySelector('.effect-level__slider');
-const effectLevelElem = overlayImageElem.querySelector('.effect-level__value');
-const sliderFieldElem = overlayImageElem.querySelector('.img-upload__effect-level');
-
-const successfulSubmission = document.querySelector('#success').content.querySelector('.success');
-const errorSubmission = document.querySelector('#error').content.querySelector('.error');
-const successButton = successfulSubmission.querySelector('.success__button');
-const errorButton = errorSubmission.querySelector('.error__button');
-
-const pristine = new Pristine(formElem, {
-  classTo: 'img-upload__field-wrapper',
-  errorClass: 'text--invalid',
-  successClass: 'text--valid',
-  errorTextParent: 'img-upload__field-wrapper',
-  errorTextTag: 'div',
-  errorTextClass: 'text--invalid__error'
-}, false);
-
-let currentEffect;
-let scaleValue = parseInt(scaleControlElem.value.replace('%', ''), 10);
-const applyEffect = (evt) => {
-  currentEffect = evt.target.value;
-  const effectModel = EFFECTS[currentEffect];
-  if (!effectModel) {
-    previewImageElem.style.filter = 'none';
-    sliderFieldElem.classList.add('hidden');
-    return;
-  }
-
-  sliderFieldElem.classList.remove('hidden');
-  const min = effectModel.min;
-  const max = effectModel.max;
-  const step = effectModel.step;
-
-  sliderElem.noUiSlider.updateOptions({
-    range: {min, max},
-    start: max,
-    step,
-  });
-  const effectsPreview = evt.target.parentNode.querySelector('.effects__preview');
-  previewImageElem.classList.add(effectsPreview.getAttribute('class').split('  ')[1]);
-
-};
-const onEffectsChangeListener = (evt) => {
-  applyEffect(evt);
-};
-const changeEffectIntensity = () => {
-  const sliderValue = sliderElem.noUiSlider.get();
-  effectLevelElem.value = sliderValue;
-  const effect = EFFECTS[currentEffect];
-  previewImageElem.style.filter = effect ? `${effect.style}(${sliderValue}${effect.unit})` : '';
-
-};
-const onZoomInListener = () => {
-  scaleValue += ZOOM_STEP;
-  scaleValue = (scaleValue > ZOOM_MAX) ? ZOOM_MAX : scaleValue;
-  scaleControlElem.value = `${scaleValue}%`;
-  previewImageElem.style.transform = `scale(${scaleValue / 100})`;
-};
-
-const onZoomOutListener = () => {
-  scaleValue -= ZOOM_STEP;
-  scaleValue = (scaleValue < ZOOM_MIN) ? ZOOM_MIN : scaleValue;
-  scaleControlElem.value = `${scaleValue}%`;
-  previewImageElem.style.transform = `scale(${scaleValue / 100})`;
-};
+const FILE_TYPES = ['gif', 'jpg', 'jpeg', 'png'];
 
 const closeOverlay = () => {
   uploadImageElem.value = '';
 
   formElem.reset();
   document.removeEventListener('keydown', onOverlayEscKeydown);
-  bnZoomOutElem.removeEventListener('click', onZoomOutListener);
-
-  bnZoomInElem.removeEventListener('click', onZoomInListener);
-  effectsListElem.removeEventListener('change', onEffectsChangeListener);
-
-  sliderElem.noUiSlider.destroy();
+  removeScaleListeners();
+  removeEffectsListeners();
   overlayImageElem.classList.add('hidden');
 
   document.body.classList.remove('modal-open');
-  pristine.destroy();
+  removePristine();
 };
-
-function onOverlayEscKeydown(evt) {
-  if (isEscKey(evt.key) && evt.target !== hashtagInputElem && evt.target !== commentInputElem) {
-    closeOverlay();
-  }
-}
 
 const disableSubmitButton = () => {
   submitButton.disabled = true;
@@ -124,116 +49,86 @@ const enableSubmitButton = () => {
 };
 
 const onCloseSuccessSendMsgClickListener = (evt) => {
-  if (evt.target === successfulSubmission) {
+  if (evt.target === successfulSubmissionElem) {
     closeSendMessages();
   }
 };
 
 const onCloseErrorSendMsgClickListener = (evt) => {
-  if (evt.target === errorSubmission) {
+  if (evt.target === errorSubmissionElem) {
     closeSendMessages();
   }
 };
-
 const onErrorSendMsgEscKeydownListener = (evt) => {
   if (isEscKey(evt.key)) {
     closeSendMessages();
   }
 };
+const onSuccess = () => {
+  closeOverlay();
+  enableSubmitButton();
+  bnSuccessElem.addEventListener('click', closeSendMessages);
+  document.addEventListener('keydown', onErrorSendMsgEscKeydownListener);
+  document.addEventListener('click', onCloseSuccessSendMsgClickListener);
+  documentBodyElem.appendChild(successfulSubmissionElem);
+};
+const onFail = () => {
+  overlayImageElem.classList.add('hidden');
+  enableSubmitButton();
+  bnErrorElem.addEventListener('click', closeSendMessages);
+  document.addEventListener('keydown', onErrorSendMsgEscKeydownListener);
+  document.addEventListener('click', onCloseErrorSendMsgClickListener);
+  documentBodyElem.appendChild(errorSubmissionElem);
+};
+
+function onOverlayEscKeydown(evt) {
+  if (isEscKey(evt.key) && evt.target !== hashtagInputElem && evt.target !== commentInputElem) {
+    closeOverlay();
+  }
+}
 
 function closeSendMessages() {
-  if (documentBody.contains(successfulSubmission)) {
-    documentBody.removeChild(successfulSubmission);
+  if (documentBodyElem.contains(successfulSubmissionElem)) {
+    documentBodyElem.removeChild(successfulSubmissionElem);
   }
-  if (documentBody.contains(errorSubmission)) {
+  if (documentBodyElem.contains(errorSubmissionElem)) {
     overlayImageElem.classList.remove('hidden');
-    documentBody.removeChild(errorSubmission);
+    documentBodyElem.removeChild(errorSubmissionElem);
   }
 
   document.removeEventListener('keydown', onErrorSendMsgEscKeydownListener);
   document.removeEventListener('click', onCloseSuccessSendMsgClickListener);
   document.removeEventListener('click', onCloseErrorSendMsgClickListener);
-  successButton.removeEventListener('click', closeSendMessages);
-  errorButton.removeEventListener('click', closeSendMessages);
+  bnSuccessElem.removeEventListener('click', closeSendMessages);
+  bnErrorElem.removeEventListener('click', closeSendMessages);
 }
-
-const onSuccess = () => {
-  closeOverlay();
-  enableSubmitButton();
-  successButton.addEventListener('click', closeSendMessages);
-  document.addEventListener('keydown', onErrorSendMsgEscKeydownListener);
-  document.addEventListener('click', onCloseSuccessSendMsgClickListener);
-  documentBody.appendChild(successfulSubmission);
-};
-const onFail = () => {
-  overlayImageElem.classList.add('hidden');
-  enableSubmitButton();
-  errorButton.addEventListener('click', closeSendMessages);
-  document.addEventListener('keydown', onErrorSendMsgEscKeydownListener);
-  document.addEventListener('click', onCloseErrorSendMsgClickListener);
-  documentBody.appendChild(errorSubmission);
-};
 
 const initUploadingFeature = () => {
   uploadImageElem.addEventListener('change', () => {
     document.addEventListener('keydown', onOverlayEscKeydown);
     bnCloseElem.addEventListener('click', closeOverlay, {once: true});
 
-    scaleValue = 100;
-    scaleControlElem.value = '100%';
-    previewImageElem.style.transform = 'scale(1)';
-    bnZoomOutElem.addEventListener('click', onZoomOutListener);
-    bnZoomInElem.addEventListener('click', onZoomInListener);
-
-    currentEffect = 'effect-none';
-    previewImageElem.classList.add('effects__preview--none');
-    effectsListElem.addEventListener('change', onEffectsChangeListener);
-
-    sliderFieldElem.classList.add('hidden');
-    noUiSlider.create(sliderElem, {
-      range: {min: 0, max: 100},
-      start: 100,
-      connect: 'lower'
-    });
-    sliderElem.noUiSlider.on('update', () => {
-      changeEffectIntensity();
-    });
+    initScaleSettings();
+    initEffectsSettings();
 
     document.body.classList.add('modal-open');
     overlayImageElem.classList.remove('hidden');
   });
+  initPristineValidations();
+  fileChooserElem.addEventListener('change', () => {
+    const file = fileChooserElem.files[0];
+    const fileName = file.name.toLowerCase();
 
-  pristine.addValidator(
-    hashtagInputElem,
-    (input) => areHashtagsRightAmount(input),
-    'Максимальное количество хештегов - 5!',
-    1
-  );
+    const matches = FILE_TYPES.some((it) => fileName.endsWith(it));
 
-  pristine.addValidator(
-    hashtagInputElem,
-    (input) => hasNoHashtagDuplicates(input),
-    'Хэштэги не должны повторяться!',
-    2
-  );
-
-  pristine.addValidator(
-    hashtagInputElem,
-    (input) => areHashtagsCorrect(input),
-    'Хэштэг может содержать только буквы и цифры. Максимальная длина хештега - 20 символов.',
-    3
-  );
-
-  pristine.addValidator(
-    commentInputElem,
-    (input) => isCommentCorrect(input),
-    `Длина комментария не должна превышать ${MAX_COMMENT_SIZE} символов`
-  );
+    if (matches) {
+      previewElem.src = URL.createObjectURL(file);
+    }
+  });
 
   formElem.addEventListener('submit', (evt) => {
-    const isValid = pristine.validate();
     evt.preventDefault();
-    if (isValid) {
+    if (isPristineValid()) {
       disableSubmitButton();
       uploadPicture(onSuccess, onFail, new FormData(evt.target));
     }
